Created Application class/ Created C# style Action event class/ Created Window class/ Setup OpenGL callback functions/ Created test level
Application class creates a window, which creates and stores the OpenGL window based on the settings parameters.

The Behaviours can be added to the Application and will updated every frame.

When an event from OpenGL is called, the OnEvent method from the Application will be called with the assosiated Event reference (the event info).

The Action class works similar as the C# Action class, where you can create Actions based on a variadic template and bind/unbind methods with the same arguments as paramaters. It also supports lambda's.